Potential Shift to Leaflet
Robin thinks we should make the switch to Leaflet, so I've added a
basemap but am struggling to add the topojsons. I think we should be
able to do this with d3 still, but I haven't figured it out yet. I
didn't delete any code from main.js but just commented it out in the
meantime.